+++ src/types/relationships.ts
@@ -1,5 +1,10 @@
 import { ExplicitRelationship } from '@jupiterone/integration-sdk-core';
 
-export interface RepoTeamRelationship extends ExplicitRelationship {
-  permission: string;
+export interface RepoAllowRelationship extends ExplicitRelationship {
+  role: string; // 'READ' | 'TRIAGE' | 'WRITE' | 'MAINTAIN' | 'ADMIN'
+  adminPermission: boolean;
+  maintainPermission: boolean;
+  pushPermission: boolean;
+  triagePermission: boolean;
+  pullPermission: boolean;
 }